Changed to proper PDP object

+++ src/main/java/ca/fourthreethreefour/settings/Settings.java
@@ -110,5 +110,5 @@ public interface Settings {
 	double DISTANCE_TOLERANCE = settingsFile.getDoubleProperty("DISTANCE_TOLERANCE", 50);
 	double TURN_TOLERANCE = settingsFile.getDoubleProperty("TURN_TOLERANCE", 5);
 	
-	byte RAMP_CHANNEL = (byte)settingsFile.getIntProperty("RAMP_CHANNEL", 0);
+	int RAMP_CHANNEL = settingsFile.getIntProperty("RAMP_CHANNEL", 0);
 }
